+++ src/index.ts
@@ -1,4 +1,5 @@
 export * from './components';
 export * from './definitions';
 export * from './helpers';
+export * from './metadata';
 export * from './module';